+++ frontgo.gemspec
@@ -34,4 +34,5 @@ Gem::Specification.new do |spec|
   spec.require_paths = ["lib"]
 
   spec.add_dependency "faraday"
+  spec.add_dependency "faraday-retry"
 end
