+++ lib/frontgo.rb
@@ -1,9 +1,6 @@
 # frozen_string_literal: true
 
 require_relative "frontgo/version"
-
-require_relative "middleware/raise_error"
-
 require_relative "frontgo/connection"
 require_relative "frontgo/order"
 require_relative "frontgo/reservation"
@@ -13,9 +10,10 @@ require_relative "frontgo/refund"
 require_relative "frontgo/terminal"
 require_relative "frontgo/credit"
 
-require_relative "frontgo/error"
-
 require "faraday"
+require "faraday/retry"
+
+require_relative "middleware/raise_error"
 
 module Frontgo
   class Error < StandardError; end
@@ -41,6 +39,14 @@ module Frontgo
       @demo = demo || false
       @connection = Faraday.new(base_url) do |conn|
         conn.headers["Authorization"] = "Bearer #{key}"
+        # We're using default settings, just adding 'retry_if" option.
+        # Originally faraday only retries for TimeoutError's, we're extending it to all ServerErrors
+        conn.request :retry, max: 2,
+          interval: 0.05,
+          interval_randomness: 0.5,
+          backoff_factor: 2,
+          retry_if: ->(_env, exception) { exception.try(:retriable?) }
+
         conn.request :json
         conn.response :json
         conn.response :frontgo_raise_error
